requirements.txt
click==8.1.7
rich==13.7.1
pyyaml==6.0.2
requests==2.32.3
pytesseract==0.3.10
Pillow==10.4.0
py7zr==0.21.0
pdfplumber==0.11.4
regex==2024.9.11
python-dotenv==1.0.1
gitpython==3.1.43
psycopg2-binary==2.9.9
pymongo==4.8.0
pymysql==1.1.1
redis==5.0.8
firebase-admin==6.6.0
boto3==1.34.162
azure-storage-blob==12.24.0
paramiko==3.5.0
cryptography==43.0.1
protobuf<5,>=4.21.0
pdfminer.six==20231228
python-docx==1.1.2
chardet==5.2.0
openpyxl==3.1.5
pandas==2.2.2
langdetect==1.0.9
phonenumberslite==8.13.43

# Optional connectors
google-cloud-storage==2.18.2
google-api-python-client==2.147.0
google-auth==2.34.0
slack-sdk==3.31.0
couchdb==1.2

openai>=1.0.0
anthropic>=0.18.0

